refactor: 컨벤션에 맞게 이름 변경
