+++ backend/src/main/java/com/metamong/mt/global/config/DataLoader.java
@@ -6,13 +6,14 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Component;
 
 import com.metamong.mt.member.model.Member;
-import com.metamong.mt.member.repository.mybatis.IMemberRepository;
+import com.metamong.mt.member.model.Role;
+import com.metamong.mt.member.repository.mybatis.MemberMapper;
 
 @Component
 public class DataLoader implements CommandLineRunner {
 
     @Autowired
-    private IMemberRepository memberMapper;
+    private MemberMapper memberMapper;
 
     @Autowired
     private PasswordEncoder passwordEncoder;
@@ -25,12 +26,12 @@ public class DataLoader implements CommandLineRunner {
 
             // 빌더 패턴을 사용하여 admin 계정 생성
             Member admin = Member.builder()
-                    .userid("admin")
+                    .userId("admin")
                     .name("Admin")
                     .password(encodedPassword)
                     .phone("000-0000-0000")
                     .email("admin@example.com")
-                    .role("ROLE_ADMIN")
+                    .role(Role.ROLE_ADMIN)
                     .postalCode("00000")
                     .address("서울시 예시구 예시동")
                     .detailAddress("상세 주소 예시")
