+++ backend/src/main/java/com/metamong/mt/member/model/Member.java
@@ -2,14 +2,22 @@ package com.metamong.mt.member.model;
 
 import java.time.LocalDate;
 
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.Id;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
+@Entity
 @Getter @Setter
 @Builder
 public class Member {
-    private String userid;
+    
+    @Id
+    private String userId;
+    
     private String name;
     private String password;
     private String phone;
@@ -18,7 +26,10 @@ public class Member {
     private String postalCode;
     private String detailAddress;
     private String address;
-    private String role;
+    
+    @Enumerated(EnumType.STRING)
+    private Role role;
+    
     private String refreshToken;
     private String businessName;
     private String businessRegistrationNumber;
