+++ backend/src/main/java/com/metamong/mt/member/model/Role.java
@@ -1,0 +1,17 @@
+package com.metamong.mt.member.model;
+
+public enum Role {
+    ROLE_USER,
+    ROLE_OWNER,
+    ROLE_ADMIN;
+    
+    private final String role;
+    
+    Role() {
+        this.role = name().substring("ROLE_".length());
+    }
+    
+    public String role() {
+        return this.role;
+    }
+}
