+++ backend/src/test/java/com/metamong/mt/BackendApplicationTests.java
@@ -1,4 +1,4 @@
-package com.metamong;
+package com.metamong.mt;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
